+++ packages/waffle-tool/cli.js
@@ -1,3 +1,3 @@
 #!/usr/bin/env node
 
-require('./waflle-tool');
+require('./waffle-tool');
