+++ packages/waffle-tool/constants/index.js
@@ -1,7 +1,9 @@
 const DEFAULT_CONFIG = '/.waffle';
 const CREATE_GENERATOR = 'create-generator';
+const COMPONENT_MODULE = 'componentModule'
 
 module.exports = {
   DEFAULT_CONFIG,
-  CREATE_GENERATOR
+  CREATE_GENERATOR,
+  COMPONENT_MODULE
 }
