+++ packages/waffle-tool/waffle-tool.js
@@ -1,18 +1,21 @@
 const {Command} = require("commander");
 const program = new Command();
 const semver = require("semver");
-const config = require("../auto-entry/create-components-entry");
-const { runGenerator } = require('../create-generator/createGenerator');
+const config = require("./auto-entry/create-components-entry");
+const { runGenerator } = require('./create-generator/createGenerator');
+const chalk = require('chalk');
 
 program
-  .version(require("../package.json").version)
+  .version(require("./package.json").version)
 
 program
   .command('create <type>')
   .alias('-c')
   .description('创建组件模板')
   .action(async function(type){
-    runGenerator({type});
+    console.log(type)
+    console.log(chalk.green('创建组件模板'));
+    await runGenerator({type});
   })
 
 program
@@ -20,6 +23,7 @@ program
   .alias('-e')
   .description('构建入口文件')
   .action(async function(){
+    console.log(chalk.green('构建入口文件'));
     await config();
   })
 
@@ -30,6 +34,4 @@ if (!semver.satisfies(process.version, ">= 8.0.0")) {
     chalk.red("✘ The generator will only work with Node v8.0.0 and up!")
   );
   process.exit(1);
-}
-
-
+};
